+++ src/components/index.js
@@ -3,3 +3,4 @@ export { default as SortableList } from './SortableList';
 export { default as LoadingButton } from './LoadingButton';
 export { default as ErrorModal } from './ErrorModal';
 export { default as PrintButton } from './PrintButton';
+export { default as PrintContent } from './PrintContent';